+++ src/algorithm/math/isPrime/index.ts
@@ -1,0 +1,16 @@
+export function isPrime(n) {
+  if (n <= 3) {
+    return n > 1;
+  }
+
+  if (n % 2 == 0 || n % 3 == 0) {
+    return false;
+  }
+
+  for (let i = 5; i * i <= n; i += 6) {
+    if (n % i == 0 || n % (i + 2) == 0) {
+      return false;
+    }
+  }
+  return true;
+}